added fake data generator to dat0r core
Added a Dat0r\Core\Runtime\Sham\DataGenerator class to create new
documents with fake data or fill existing documents with fake data. See
the tests or doc comments for more information about usage.

This change introduces fzaninotto/faker as a dependency for Dat0r. The lib
is used to create fake data for the documents.