first step in refactoring the Sham\DataGenerator
The DataGenerator grew a bit large and the fake data generation was a big
switch case block for each known and supported field type. The single
strategies to generate fake data for a specific type have now been
extracted into methods.

The TODO item was added as it seems, that references and aggregate do not
trigger change events upon changes to those fields.