explicitely allow callables that are not closures
Instead of implicitely calling `$closure()` within the code for getting
the value to set, `call_user_func` is now being used. As of now a simple
callable like `array($this, 'methodName')` is already functioning, but it
was not in the API docs (as executing `$array()` is not common knowledge).

You can now use static calls:

DataGenerator::fill($this->document, array(
    DataGenerator::OPTION_FIELD_VALUES => array(
        'fieldname' => 'Dat0r\\Tests\\Core\\Runtime\\Sham\\DataGeneratorTest::getStaticTrololo'
    )
));

or array syntax for method calls:

DataGenerator::fill($this->document, array(
    DataGenerator::OPTION_FIELD_VALUES => array(
        'fieldname' => array($this, 'getTrololo')
    )
));