refactored *ListAttributes and changes to
Boolean and Integer to be consistent when handling values. Add BooleanList
and adjust boolean casted KeyValueList to handle boolean values the same
as the BooleanAttribute.

Introduce getNativeType and getValueType methods on the
ValueHolderInterface to reflect the types used as return values.

Adjust handling of default an null values for consistency.

Add some more tests for e.g. the new MIN_COUNT and MAX_COUNT options for
*ListAttributes.