remove ini_get precision handling as that setting
is only taken into account when float values are stringified in php and
then are used for the overall number of digits instead of only the
fractional part of the number. Pro tip: to format floats as strings use
sprintf('%0.8f', $float) or similar as sprintf uses whatever precision is
wanted instead of interpreting the 'precision' ini setting.